fix(board): one REAL timestamp took GET /api/board down fleet-wide (AF-317 fallout)
I caused this and it was live for roughly twenty minutes.

`queue_disposition`'s update branch wrote `updated` with `now_f64()`. SQLite is
dynamically typed, so an f64 into an INTEGER column stores REAL in that row, and
rusqlite's FromSql is strict per storage type — so `issue_from_row` failed the
WHOLE list read:

    {"error":"Invalid column type Real at index: 6, name: updated"}

12,959 correct rows made unreadable by 3 wrong ones, and `GET /api/board`
returned that to every session in the fleet. The three were the disposition
cards the job filed on its first tick and rewrote on its second, which is why
the outage arrived a tick AFTER the deploy rather than with it.

Data repaired in place (CAST to INTEGER on the 3 rows); the board is serving
2,009 items again.

Two fixes, because the writer fix alone would leave the fragility:

1. The job writes `now_f64() as i64`, like every other writer.
2. `updated` is read through a type-erased `ts_i64` at BOTH call sites. A list
   read over ~13,000 rows must not be all-or-nothing on the storage type of a
   single cell. The precedent was already in the same function — `last_verified_at`
   has carried exactly this treatment for the Python-era TEXT case — and this
   column did not have it.

Regression test reproduces the incident: a REAL and an INTEGER row in one table,
both must come back, the INTEGER one exact so a reader that zeroed everything
would fail. It also asserts that a strict i64 read of the REAL cell still errors,
so the fixture is known to reproduce the hazard rather than a hypothetical.

Also finishes 22546f0d, which corrected the AF-317 numbers in the create-path
refusal but missed the PATCH-path copy: it was still telling every lane
"the median todo card is 28.8 days old", which is the archived-contaminated
figure. Both refusal strings rebuilt (the python edits that made those
corrections had also collapsed their line continuations into runs of spaces).

Suite 1601 passed, 0 failed; clippy -D warnings clean.

Amux-Session: amux-frustrations
Amux-Conversation: 051edaf5-6e80-47b3-b4cf-57e427c0e713

+++ crates/amux-server/src/runtime_jobs/queue_disposition.rs
@@ -200,9 +200,19 @@ pub async fn tick(state: AppState) -> (usize, usize, usize) {
                 let _ = state
                     .store
                     .write_async(move |conn| {
+                        // INTEGER, not f64. `updated` is an INTEGER column and
+                        // every other writer stores it as one; SQLite is
+                        // dynamically typed, so a single f64 here silently
+                        // stores REAL in that row and `issue_from_row` then
+                        // fails the WHOLE list read with "Invalid column type
+                        // Real at index: 6". Three rows written by this job
+                        // took `GET /api/board` down fleet-wide for ~20 minutes
+                        // on 2026-08-30. The reader is hardened too (see
+                        // `issue_from_row`); this is the half that stops
+                        // producing the bad value in the first place.
                         conn.execute(
                             "UPDATE issues SET title=?1, \"desc\"=?2, updated=?3 WHERE id=?4",
-                            rusqlite::params![t, d, crate::config::now_f64(), id],
+                            rusqlite::params![t, d, crate::config::now_f64() as i64, id],
                         )?;
                         Ok(crate::db::WriteOutcome { applied: true, events: vec![] })
                     })
